requirements.txt
fastapi[standard]
aiohttp
python-dotenv
httpx
astrapy